separate manual testing code from extension code in webview-ui
