+++ webview-ui/index.html
@@ -8,6 +8,6 @@
   </head>
   <body>
     <div id="root" data-contentid="" data-initialstate="{}"></div>
-    <script type="module" src="/src/main.tsx"></script>
+    <script type="module" src="/src/manualTest/main.tsx"></script>
   </body>
 </html>
